# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Application API Documentation",
    "version": "1.0.0",
    "description": "API documentation for the generated application",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "https://your-production-url.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
  "security": [],
  "paths": {
...
    "/api/coupons/{id}": {
      "get": {
        "summary": "Get coupon by ID",
        "tags": [
          "Coupons"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Coupon details"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/api/coupons": {
      "get": {
        "summary": "Get all coupons",
        "tags": [
          "Coupons"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of coupons"
          }
        }
      },
      "post": {
        "summary": "Create new coupon",
        "tags": [
          "Coupons"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Coupon created"
          }
        }
      }
    },
...
    "/api/export": {
      "post": {
        "summary": "Export data to CSV",
        "tags": [
          "Data"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "workflow": {
                    "type": "string",
                    "enum": [
                      "seo",
                      "websites",
                      "coupons"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "CSV data"
          }
        }
      }
    },
    "/api/import": {
      "post": {
        "summary": "Import data from CSV",
        "tags": [
          "Data"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "workflow": {
                    "type": "string",
                    "enum": [
                      "seo",
                      "websites",
                      "coupons"
                    ]
                  },
                  "csvData": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Data imported successfully"
          }
        }
      }
    },
...
  },
...
}